task_01/requirements.txt
pandas
openai
python-dotenv
ipykernel
streamlit